int32 -> sd-int32 and rewrite.

+++ SYSDEPS/modules/sd-int32/try_stdint.c
@@ -1,8 +1,9 @@
 #include <stdint.h>
 
-int main(void)
+int
+main(void)
 {
-  int64_t n;
+  int32_t n;
   n = 128; /* for pedantic compilers */
   return 0;
 }
